mysql: small privileges refactor (#1198)
* mysql: extract privs out


* mysql: add priv.String()


* mysql: add privs.Has()


* mysql: add p.SetString() and p.ColumnString()


* mysql: add NewPrivFromSetEnum/Column


* *: add license


* *: complete the license
